Move all tests to single package
