+++ tests/api-controller-acceptance-tests/apicontroller/component_test.go
@@ -1,4 +1,4 @@
-package component
+package apicontroller
 
 import (
 	"fmt"
