+++ tests/api-controller-acceptance-tests/apicontroller/fixture.go
@@ -1,4 +1,4 @@
-package acceptance
+package apicontroller
 
 import (
 	"fmt"
