+++ tests/api-controller-acceptance-tests/apicontroller/integration_test.go
@@ -1,4 +1,4 @@
-package acceptance
+package apicontroller
 
 import (
 	"context"
